refactor(price estimator): Rename for clarity PE-126
