--- src/utils/price_estimator.ts
@@ -1,4 +1,0 @@
-export interface PriceEstimator {
-	getARPriceForByteCount: (byteCount: number) => Promise<number>;
-	getByteCountForAR: (arPrice: number) => Promise<number>;
-}
